backend/services/upload_service.py
import asyncio
import logging
from typing import Dict, Any, Optional
from sqlalchemy.orm import Session
from backend.database import SessionLocal
from backend.models import Account, PlatformType
from backend.config import settings

logger = logging.getLogger(__name__)

class UploadService:
    """
    Service for handling video uploads to various platforms.
    
    IMPORTANT: This is currently a SIMULATED service for demonstration purposes.
    
    For REAL uploads to work, you need to implement:
    1. OAuth 2.0 authentication flow for each platform
    2. Platform-specific API integrations:
       - YouTube: Google API Client (youtube.videos().insert())
       - TikTok: TikTok Content Posting API
       - Douyin: Douyin Open Platform API
    3. Secure token storage and refresh mechanisms
    4. Video metadata handling (title, description, tags, privacy settings)
    5. Thumbnail upload support
    6. Error handling for platform-specific rate limits and restrictions
    
    The Account model already has fields for OAuth tokens:
    - access_token: OAuth access token for API authentication
    - refresh_token: Token to refresh expired access tokens
    
    Users must authenticate their accounts (similar to channel scanning) before uploads can work.
    """
    
    def __init__(self):
        pass
        
    async def upload_video(self, video_path: str, account_id: int, platform: str, progress_callback=None) -> bool:
        """
        Upload a video to the specified platform using the given account.
        """
        db = SessionLocal()
        try:
            # Validate account exists
            account = db.query(Account).filter(Account.id == account_id).first()
            if not account:
                if progress_callback:
                    await progress_callback(f"Account ID {account_id} not found", 0)
                return False
            
            # Validate platform match
            if account.platform != platform:
                if progress_callback:
                    await progress_callback(f"Account platform mismatch: expected {platform}, got {account.platform}", 0)
                return False
            
            # Check for authentication credentials
            if not account.access_token:
                if progress_callback:
                    await progress_callback(
                        f"Account '{account.username}' is not authenticated. Please link your {platform} account with OAuth credentials.",
                        0
                    )
                logger.warning(f"Upload attempted with unauthenticated account {account_id}")
                return False
                
            # ============================================================
            # REAL UPLOAD IMPLEMENTATION
            # ============================================================
            
            if platform == PlatformType.YOUTUBE:
                return await self._upload_to_youtube(video_path, account, progress_callback)
            
            # Fallback for other platforms (Simulation)
            else:
                return await self._simulate_upload(video_path, account, platform, progress_callback)
            
        except Exception as e:
            logger.error(f"Upload failed: {e}")
            if progress_callback:
                await progress_callback(f"Upload failed: {str(e)}", 0)
            return False
        finally:
            db.close()

    async def _upload_to_youtube(self, video_path: str, account: Account, progress_callback=None) -> bool:
        """Upload video to YouTube using Google API"""
        import google.oauth2.credentials
        import googleapiclient.discovery
        from googleapiclient.http import MediaFileUpload
        import os

        try:
            if progress_callback:
                await progress_callback(f"Initializing YouTube upload for {account.username}...", 0)

            # Create credentials object
            creds = google.oauth2.credentials.Credentials(
                token=account.access_token,
                refresh_token=account.refresh_token,
                token_uri="https://oauth2.googleapis.com/token",
                client_id=settings.YOUTUBE_CLIENT_ID,
                client_secret=settings.YOUTUBE_CLIENT_SECRET,
            )

            # Build service
            service = googleapiclient.discovery.build('youtube', 'v3', credentials=creds)

            # Prepare metadata
            filename = os.path.basename(video_path)
            title = os.path.splitext(filename)[0].replace('_', ' ').title()
            
            body = {
                'snippet': {
                    'title': title,
                    'description': f"Uploaded via VidFlow: {title}",
                    'tags': ['vidflow', 'auto-upload'],
                    'categoryId': '22'  # People & Blogs
                },
                'status': {
                    'privacyStatus': 'private',  # Default to private for safety
                    'selfDeclaredMadeForKids': False
                }
            }

            # Create media file object (resumable for progress tracking)
            media = MediaFileUpload(
                video_path, 
                mimetype='video/mp4',
                resumable=True,
                chunksize=1024*1024  # 1MB chunks
            )

            # Create insert request
            request = service.videos().insert(
                part=','.join(body.keys()),
                body=body,
                media_body=media
            )

            # Execute upload with progress tracking
            response = None
            while response is None:
                status, response = request.next_chunk()
                if status:
                    progress = int(status.progress() * 100)
                    if progress_callback:
                        await progress_callback(f"Uploading to YouTube... {progress}%", progress)
                    # Allow other tasks to run
                    await asyncio.sleep(0.1)

            if response and 'id' in response:
                video_id = response['id']
                if progress_callback:
                    await progress_callback(f"Upload complete! Video ID: {video_id}", 100)
                logger.info(f"YouTube upload successful. Video ID: {video_id}")
                return True
            else:
                raise Exception("Upload completed but no video ID returned")

        except Exception as e:
            logger.error(f"YouTube upload error: {e}")
            if progress_callback:
                await progress_callback(f"YouTube upload failed: {str(e)}", 0)
            return False

    async def _simulate_upload(self, video_path: str, account: Account, platform: str, progress_callback=None) -> bool:
        """Simulated upload for other platforms"""
        filename = video_path.split('/')[-1].split('\\')[-1]
        
        if progress_callback:
            await progress_callback(f"[SIMULATED] Starting upload of {filename} to {platform} ({account.username})", 0)
        
        await asyncio.sleep(1)
        
        chunks = 5
        for i in range(chunks):
            progress = int((i + 1) / chunks * 100)
            if progress_callback:
                await progress_callback(f"[SIMULATED] Uploading... {progress}%", progress)
            await asyncio.sleep(1)
            
        if progress_callback:
            await progress_callback(f"[SIMULATED] Upload complete! Video is live on {platform}.", 100)
            
        return True

# Global instance
upload_service = UploadService()
