backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
yt-dlp>=2024.1.0
ffmpeg-python>=0.2.0
deep-translator>=1.11.4
websockets>=12.0
python-multipart>=0.0.6
aiofiles>=23.2.1
httpx>=0.26.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
alembic>=1.13.1


google-auth-oauthlib>=1.2.0
google-api-python-client>=2.118.0
